MDC implemented, Test pending, get wallet and user info implemented

+++ ewallet/userService/src/main/java/com/user/repository/UserRepository.java
@@ -3,6 +3,6 @@ package com.user.repository;
 import com.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer>
+public interface UserRepository extends JpaRepository<User, Long>
 {
 }
